computer vision

+++ train_cv.py
@@ -28,9 +28,9 @@ for file in os.listdir(datasetdir):
         y1 = float(annotation_dict[file]['y1']) / shape[0]
         x2 = float(annotation_dict[file]['x2']) / shape[1]
         y2 = float(annotation_dict[file]['y2']) / shape[0]
-        y.append(np.array([x1, y1, x2, y2]))
+        y.append(np.array([x1, y1, x2, y2]).astype(np.float32))
     else:
-        y.append(np.array([0, 0, 0, 0]))
+        y.append(np.array([0, 0, 0, 0].astype(np.float32)))
 x = np.array(x).astype(np.float32)
 y = np.array(y).astype(np.float32)
